+++ modules/AWSPowerShell/Cmdlets/ConfigService/Basic/Add-CFGResourceTag-Cmdlet.cs
@@ -30,8 +30,8 @@ namespace Amazon.PowerShell.Cmdlets.CFG
     /// <summary>
     /// Associates the specified tags to a resource with the specified resourceArn. If existing
     /// tags on a resource are not specified in the request parameters, they are not changed.
-    /// When a resource is deleted, the tags associated with that resource are deleted as
-    /// well.
+    /// If existing tags are specified, however, then their values will be updated. When a
+    /// resource is deleted, the tags associated with that resource are deleted as well.
     /// </summary>
     [Cmdlet("Add", "CFGResourceTag", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("None")]