+++ modules/AWSPowerShell/Cmdlets/ConfigService/Basic/Get-CFGComplianceByResource-Cmdlet.cs
@@ -67,7 +67,7 @@ namespace Amazon.PowerShell.Cmdlets.CFG
         #region Parameter ComplianceType
         /// <summary>
         /// <para>
-        /// <para>Filters the results by compliance.</para><para>The allowed values are <code>COMPLIANT</code>, <code>NON_COMPLIANT</code>, and <code>INSUFFICIENT_DATA</code>.</para>
+        /// <para>Filters the results by compliance.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]