+++ modules/AWSPowerShell/Cmdlets/ConfigService/Basic/Start-CFGResourceEvaluation-Cmdlet.cs
@@ -36,7 +36,17 @@ namespace Amazon.PowerShell.Cmdlets.CFG
     /// 
     ///  <note><para>
     /// Ensure you have the <code>cloudformation:DescribeType</code> role setup to validate
-    /// the resource type schema. 
+    /// the resource type schema.
+    /// </para><para>
+    /// You can find the <a href="https://docs.aws.amazon.com/cloudformation-cli/latest/userguide/resource-type-schema.html">Resource
+    /// type schema</a> in "<i>Amazon Web Services public extensions</i>" within the CloudFormation
+    /// registry or with the following CLI commmand: <code>aws cloudformation describe-type
+    /// --type-name "AWS::S3::Bucket" --type RESOURCE</code>.
+    /// </para><para>
+    /// For more information, see <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/registry.html#registry-view">Managing
+    /// extensions through the CloudFormation registry</a> and <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-template-resource-type-ref.html">Amazon
+    /// Web Services resource and property types reference</a> in the CloudFormation User
+    /// Guide.
     /// </para></note>
     /// </summary>
     [Cmdlet("Start", "CFGResourceEvaluation", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
@@ -62,7 +72,8 @@ namespace Amazon.PowerShell.Cmdlets.CFG
         #region Parameter EvaluationMode
         /// <summary>
         /// <para>
-        /// <para>The mode of an evaluation. The valid value for this API is <code>Proactive</code>.</para>
+        /// <para>The mode of an evaluation. The valid values for this API are <code>DETECTIVE</code>
+        /// and <code>PROACTIVE</code>.</para>
         /// </para>
         /// </summary>
         #if !MODULAR
@@ -107,7 +118,13 @@ namespace Amazon.PowerShell.Cmdlets.CFG
         #region Parameter ResourceDetails_ResourceConfigurationSchemaType
         /// <summary>
         /// <para>
-        /// <para>The schema type of the resource configuration.</para>
+        /// <para>The schema type of the resource configuration.</para><note><para>You can find the <a href="https://docs.aws.amazon.com/cloudformation-cli/latest/userguide/resource-type-schema.html">Resource
+        /// type schema</a>, or <code>CFN_RESOURCE_SCHEMA</code>, in "<i>Amazon Web Services public
+        /// extensions</i>" within the CloudFormation registry or with the following CLI commmand:
+        /// <code>aws cloudformation describe-type --type-name "AWS::S3::Bucket" --type RESOURCE</code>.</para><para>For more information, see <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/registry.html#registry-view">Managing
+        /// extensions through the CloudFormation registry</a> and <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-template-resource-type-ref.html">Amazon
+        /// Web Services resource and property types reference</a> in the CloudFormation User
+        /// Guide.</para></note>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]