+++ modules/AWSPowerShell/Cmdlets/ConfigService/Basic/Write-CFGAggregationAuthorization-Cmdlet.cs
@@ -29,7 +29,14 @@ namespace Amazon.PowerShell.Cmdlets.CFG
 {
     /// <summary>
     /// Authorizes the aggregator account and region to collect data from the source account
-    /// and region.
+    /// and region. 
+    /// 
+    ///  <note><para><code>PutAggregationAuthorization</code> is an idempotent API. Subsequent requests
+    /// won’t create a duplicate resource if one was already created. If a following request
+    /// has different <code>tags</code> values, Config will ignore these differences and treat
+    /// it as an idempotent request of the previous. In this case, <code>tags</code> will
+    /// not be updated, even if they are different.
+    /// </para></note>
     /// </summary>
     [Cmdlet("Write", "CFGAggregationAuthorization", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.ConfigService.Model.AggregationAuthorization")]