+++ modules/AWSPowerShell/Cmdlets/ConfigService/Basic/Write-CFGStoredQuery-Cmdlet.cs
@@ -32,6 +32,13 @@ namespace Amazon.PowerShell.Cmdlets.CFG
     /// be unique for a single Amazon Web Services account and a single Amazon Web Services
     /// Region. You can create upto 300 queries in a single Amazon Web Services account and
     /// a single Amazon Web Services Region.
+    /// 
+    ///  <note><para><code>PutStoredQuery</code> is an idempotent API. Subsequent requests won’t create
+    /// a duplicate resource if one was already created. If a following request has different
+    /// <code>tags</code> values, Config will ignore these differences and treat it as an
+    /// idempotent request of the previous. In this case, <code>tags</code> will not be updated,
+    /// even if they are different.
+    /// </para></note>
     /// </summary>
     [Cmdlet("Write", "CFGStoredQuery", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("System.String")]